Gil - seperated the math from the pose estimator
