+++ src/main/java/frc/robot/poseestimator/PoseEstimatorMath.java
@@ -3,127 +3,45 @@ package frc.robot.poseestimator;
 import edu.wpi.first.math.Matrix;
 import edu.wpi.first.math.Nat;
 import edu.wpi.first.math.VecBuilder;
-import edu.wpi.first.math.Vector;
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.geometry.Transform2d;
 import edu.wpi.first.math.geometry.Twist2d;
-import edu.wpi.first.math.interpolation.TimeInterpolatableBuffer;
-import edu.wpi.first.math.kinematics.SwerveDriveKinematics;
-import edu.wpi.first.math.kinematics.SwerveDriveWheelPositions;
-import edu.wpi.first.math.kinematics.SwerveModulePosition;
 import edu.wpi.first.math.numbers.N1;
 import edu.wpi.first.math.numbers.N3;
-import frc.robot.poseestimator.observations.Observation;
 import frc.robot.poseestimator.observations.OdometryObservation;
 import frc.robot.poseestimator.observations.VisionObservation;
-
-import java.util.NoSuchElementException;
 import java.util.Optional;
 
-public class PoseEstimatorMath implements IPoseEstimator{
+public class PoseEstimatorMath {
 
-    private Pose2d odometryPose;
-    private Pose2d estimatedPose;
-    private final TimeInterpolatableBuffer<Pose2d> poseBuffer;
-    private Matrix<N3, N1> standardDeviations;
-    private final SwerveDriveKinematics kinematics;
-    private SwerveDriveWheelPositions lastWheelPositions;
-    private Rotation2d lastGyroAngle;
-    private boolean isFirstOdometryUpdate;
-    private VisionObservation lastVisionObservation;
-
-    public PoseEstimatorMath() {
-        odometryPose = new Pose2d();
-        estimatedPose = new Pose2d();
-        poseBuffer = TimeInterpolatableBuffer.createBuffer(PoseEstimatorConstants.POSE_BUFFER_SIZE_SECONDS);
-        standardDeviations = new Matrix<>(Nat.N3(), Nat.N1());
-        lastWheelPositions = new SwerveDriveWheelPositions(
-                new SwerveModulePosition[] {
-                        new SwerveModulePosition(),
-                        new SwerveModulePosition(),
-                        new SwerveModulePosition(),
-                        new SwerveModulePosition()
-                });
-        lastGyroAngle = new Rotation2d();
-        isFirstOdometryUpdate = true;
-//        kinematics = SwerveConstants.KINEMATICS;
-        kinematics = null; //todo - fix when SwerveConstants are made;
-        for (int i = 0; i < 3; ++i) {
-            standardDeviations.set(i, 0, Math.pow(PoseEstimatorConstants.ODOMETRY_STANDARD_DEVIATIONS.get(i, 0), 2));
-        }
-    }
-
-    public void addOdometryObservation(OdometryObservation observation) {
-        setInitialValuesAtStart();
-        Twist2d twist = kinematics.toTwist2d(lastWheelPositions, observation.getWheelPositions());
-        lastWheelPositions = observation.getWheelPositions();
+    public static Twist2d addGyroToTwistCalculations(OdometryObservation observation, Twist2d twist, Rotation2d lastGyroAngle) {
         boolean isGyroConnected = observation.getGyroAngle() != null;
         if (isGyroConnected) {
-            twist = updateDeltaTheta(twist,observation);
+            twist = updateDeltaTheta(twist, observation, lastGyroAngle);
         }
-        odometryPose = odometryPose.exp(twist);
-        poseBuffer.addSample(observation.getTimestamp(), odometryPose);
-        estimatedPose = estimatedPose.exp(twist);
+        return twist;
     }
 
-    private void setInitialValuesAtStart() {
-        if (isFirstOdometryUpdate) {
-//            lastWheelPositions = SWERVE.getSwerveWheelPositions(0);todo - fix when swerve is made
-//            lastGyroAngle = SWERVE.getOdometryYawUpdates()[0];
-            isFirstOdometryUpdate = false;
-        }
+    public static Twist2d updateDeltaTheta(Twist2d twist, OdometryObservation observation, Rotation2d lastGyroAngle) {
+        return new Twist2d(
+                twist.dx,
+                twist.dy,
+                observation.getGyroAngle().minus(lastGyroAngle).getRadians()
+        );
     }
 
-    private Twist2d updateDeltaTheta(Twist2d twist, OdometryObservation observation) {
-        lastGyroAngle = observation.getGyroAngle();
-        return new Twist2d(twist.dx, twist.dy, observation.getGyroAngle().minus(lastGyroAngle).getRadians());
-    }
-
-    private Transform2d useKalmanOnTransform(VisionObservation observation, Pose2d estimateAtTime) {
-        double[] squaredVisionMatrix = getSquaredVisionMatrix(observation);
-        Matrix<N3, N3> visionK = kalmanFilterAlgorithm(squaredVisionMatrix);
-        Transform2d diffFromOdometry = new Transform2d(estimateAtTime, observation.getVisionPose());
-        return scaleDiffFromKalman(diffFromOdometry, visionK);
-    }
-
-    public void addVisionObservation(VisionObservation observation) {
-        lastVisionObservation = observation;
-        if(isObservationTooOld(observation)) {
-            return;
-        }
-        Optional<Pose2d> sample = poseBuffer.getSample(observation.getTimestamp());
-        if (sample.isEmpty()) {
-            return;
-        }
-        Transform2d sampleToOdometryTransform = new Transform2d(sample.get(), odometryPose);
-        Transform2d odometryToSampleTransform = new Transform2d(odometryPose, sample.get());
-        Pose2d estimateAtTime = estimatedPose.plus(odometryToSampleTransform);
-        estimatedPose = estimateAtTime
-                .plus(useKalmanOnTransform(observation,estimateAtTime))
-                .plus(sampleToOdometryTransform);
-    }
-
-    private boolean isObservationTooOld(Observation observation) {
-        try {
-            return poseBuffer.getInternalBuffer().lastKey() - PoseEstimatorConstants.POSE_BUFFER_SIZE_SECONDS > observation.getTimestamp();
-        }
-        catch (NoSuchElementException ex) {
-            return true;
-        }
-    }
-
-    private double[] getSquaredVisionMatrix(VisionObservation observation) {
+    public static double[] getSquaredVisionMatrix(VisionObservation observation) {
         int rows = observation.getStandardDeviations().getNumRows();
         double[] r = new double[rows];
         for (int row = 0; row < rows; ++row) {
             double standardDeviationInRow = observation.getStandardDeviations().get(row, 0);
-            r[row] = Math.pow(standardDeviationInRow,2);
+            r[row] = Math.pow(standardDeviationInRow,PoseEstimatorConstants.KALMAN_EXPONENT);
         }
         return r;
     }
 
-    private Matrix<N3, N3> kalmanFilterAlgorithm(double[] squaredVisionMatrix) {
+    public static Matrix<N3, N3> kalmanFilterAlgorithm(double[] squaredVisionMatrix, Matrix<N3, N1> standardDeviations) {
         Matrix<N3, N3> visionK = new Matrix<>(Nat.N3(), Nat.N3());
         for (int row = 0; row < visionK.getNumRows(); ++row) {
             double stdDev = standardDeviations.get(row, 0);
@@ -137,7 +55,14 @@ public class PoseEstimatorMath implements IPoseEstimator{
         return visionK;
     }
 
-    private Transform2d scaleDiffFromKalman(Transform2d diffFromOdometry, Matrix<N3, N3> visionK) {
+    public static Transform2d useKalmanOnTransform(VisionObservation observation, Pose2d estimateAtTime, Matrix<N3, N1> standardDeviations) {
+        double[] squaredVisionMatrix = PoseEstimatorMath.getSquaredVisionMatrix(observation);
+        Matrix<N3, N3> visionK = PoseEstimatorMath.kalmanFilterAlgorithm(squaredVisionMatrix, standardDeviations);
+        Transform2d diffFromOdometry = new Transform2d(estimateAtTime, observation.getVisionPose());
+        return scaleDiffFromKalman(diffFromOdometry, visionK);
+    }
+
+    public static Transform2d scaleDiffFromKalman(Transform2d diffFromOdometry, Matrix<N3, N3> visionK) {
         Matrix<N3, N1> kTimesDiff = visionK.times(
                 VecBuilder.fill(diffFromOdometry.getX(), diffFromOdometry.getY(), diffFromOdometry.getRotation().getRadians())
         );
@@ -148,60 +73,19 @@ public class PoseEstimatorMath implements IPoseEstimator{
         );
     }
 
-    @Override
-    public void resetPose(Pose2d initialPose) {
-        estimatedPose = initialPose;
-        lastGyroAngle = initialPose.getRotation();
-        odometryPose = initialPose;
-        poseBuffer.clear();
-    }
-
-    @Override
-    public Pose2d getEstimatedPose() {
-        return estimatedPose;
-    }
-
-    public void setOdometryStandardDeviations(double x, double y, double rotation) {
-        Vector<N3> newStandardDeviations = VecBuilder.fill(x, y, rotation);
-        for (int i = 0; i < newStandardDeviations.getNumRows(); ++i) {
-            standardDeviations.set(i, 0, Math.pow(newStandardDeviations.get(i, 0), 2));
-        }
-    }
-
-    @Override
-    public void updateOdometry(OdometryObservation odometryObservation) {
-        addOdometryObservation(odometryObservation);
-    }
-
-    @Override
-    public void resetOdometry(SwerveDriveWheelPositions wheelPositions, Rotation2d gyroAngle, Pose2d pose) {
-        this.lastWheelPositions = wheelPositions;
-        this.lastGyroAngle = gyroAngle;
-        this.odometryPose = pose;
-        poseBuffer.clear();
-    }
-
-    @Override
-    public Pose2d getOdometryPose() {
-        return odometryPose;
-    }
-
-    @Override
-    public void updateVision(VisionObservation visionObservation) {
-        addVisionObservation(visionObservation);
-    }
-
-    @Override
-    public Pose2d getVisionPose() {
-        if(lastVisionObservation != null) {
-            return lastVisionObservation.getVisionPose();
-        }
-        return null;
-    }
-
-    @Override
-    public void setStandardDeviations(Matrix<N3, N1> standardDeviations) {
-        this.standardDeviations = standardDeviations;
+    public static Pose2d combineVisionObservationAndOdometrySample(
+            Optional<Pose2d> sample,
+            VisionObservation observation,
+            Pose2d estimatedPose,
+            Pose2d odometryPose,
+            Matrix<N3, N1> standardDeviations
+    ) {
+        Transform2d sampleToOdometryTransform = new Transform2d(sample.get(), odometryPose);
+        Transform2d odometryToSampleTransform = new Transform2d(odometryPose, sample.get());
+        Pose2d estimateAtTime = estimatedPose.plus(odometryToSampleTransform);
+        return estimateAtTime
+                .plus(PoseEstimatorMath.useKalmanOnTransform(observation, estimateAtTime, standardDeviations))
+                .plus(sampleToOdometryTransform);
     }
 
 }
